Multiple users and user pages

+++ app/profile/page.jsx
@@ -32,11 +32,20 @@ const ProfilePage = () => {
     router.push(`/update-prompt?id=${post._id}`);
   };
   const handleDelete = async (post) => {
-    const res = await fetch(`/api/prompt/${post._id}`, {
-      method: "DELETE",
-    });
-    if(res.ok){
-      router.push("/");
+    const hasConfirmed = confirm(
+      "Are you sure you want to delete this prompt?"
+    );
+    if (hasConfirmed) {
+      try {
+        const res = await fetch(`/api/prompt/${post._id.toString()}`, {
+          method: "DELETE",
+        });
+
+        const filteredPosts = posts.filter((p)=>p._id !== post._id);
+        setPosts(filteredPosts);
+      } catch (err) {
+        console.log(err);
+      }
     }
   };
